% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{parmesan_watch}
\alias{parmesan_watch}
\title{#' @export
parmesan_inputs_init <- function(parmesan = NULL, value = NA){
  ids <- parmesan_input_ids(parmesan = parmesan)
  lapply(input_ids, function(i) NA)
}}
\usage{
parmesan_watch(input, parmesan = NULL)
}
\description{
#' @export
parmesan_inputs_init <- function(parmesan = NULL, value = NA){
  ids <- parmesan_input_ids(parmesan = parmesan)
  lapply(input_ids, function(i) NA)
}
}
